Cleanup model.py by creating modules.py, remove "huge" and "mega"
